pyproject.toml
[tool.poetry]
name = "idoven-api"
version = "0.0.1"
description = "A challenge "
authors = ["Marc Vivancos <me@marcvivancos.com>"]
readme = "README.md"
packages = [{include = "ecg"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.95.0"}
pydantic = "^1.10.6"
uvicorn = "^0.21.1"
databases = "^0.7.0"
asyncpg = "^0.27.0"
pyjwt = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
click = "^8.1.3"
ipdb = "^0.13.13"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pre-commit = "^3.2.0"
black = "^23.1.0"
ipython = "^8.11.0"
ruff = "^0.0.257"
httpx = "^0.23.3"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
aiosqlite = "^0.18.0"
async-asgi-testclient = "^1.4.11"
async-generator = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 89
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
show_error_codes = true
disable_error_code = "annotation-unchecked"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]
